+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+- Remove dependency on BeautifulSoup by finding the title during rendering
+- Add support for rendering image attachments
+
 ## 0.2.2 - 2019-10-22
 - Bump depdendencies